NCBC-4046 Parse config improvement
Motivation
==========
When multiple servers share a host name (or we use the fit proxy),
we raise an exception when trying to parse the config.  That's bad.

Modification
============
We now use the nodesExt hostname - with the port - and match that
to the hostname and kv port in the vBucketMap's server list, when
constructing a map of server groups -> vbucket indexes.   Very small
change, which only matters when multiple servers share the same
hostname (admittedly rare, but common enough in our tests).

Results
=======
Fit tests that use proxy now pass.

Change-Id: I882e7407ea2cc87aad923112fb93fd7422267487
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/231407
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>