NCBC-3956 Implement GetMulti for transactions
Motivation
==========
The extGetMulti extension brings the transactions in .NET to pairity
or better with other sdks.

Modification
============
Added a GetMulti and GetMultiReplicaFromPreferredServerGroup to the
AttemptContext.   They take the appropriate specs and options, which
are also new.  The logic is all contained in the GetMultiManager
class.

Note that we added a new ReadPreference, SelectedServerGroupWithFallback,
which we implemented to fall back to requesting a doc across all the
replicas.   This is perhaps more useful as an internal preference, but
it is public for now.

Results
The 88 or so extGetMulti FIT tests all pass.

Change-Id: I72f0829b6de358b09a1e76916de22895a475d7dc
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/231592
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>