NCBC-4034: Operational SDK prevent connection to Analytics 2.0 Cluster
Motivation
----------
The operational SDK should throw a CouchbaseException when Analytics Queries are attempted if the remote server is an EA cluster.

Changes
-------
Upon retrieving a random node with the Analytics service, we check if the Product (json `prod`) of the GlobalConfig is not null and == "analytics". If so, we throw.
Added an internal bool in Experiments of ClusterOptions to bypass this behaviour.

Change-Id: Ib2363d25ee255602068aed841bcde29ea1291955
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/231714
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>