NCBC-3974: Transaction fit test cleanup
Motivation
==========
Fix of minor issues in Fit tests for transactions

Modification
============
Cleanup was not using the CAS correctly in certain circumstances,
so we now use that cas as needed.   Also - we occasionally will see
a CouchbaseException when a document cannot be written to a collection
due to permissions, instead of an UnambiguousTimeoutException.   We
now have a small heuristic - if the elapsed time is more than 95% of
the operation timeout, we consider it a timeout and not a operation
that was canceled.

Results
=======
Several cleanup and RbacOnCollections tests now pass consistently.

Change-Id: I1305869882d893151792e63fd5a997857b5c7766
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/232331
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>