NCBC-3974 Transactons Fit test issues.
Motivation
==========
Clean up last few issues with fit tests for transactions.

Modifications
=============
Ambiguous error handling was slightly off for committing an insert
or replace.  Simple fix.   Also if the internal rollbackAsync is
called, the tests expect we raise an exception.  Since this is an
internal call and only used by the tests, we now do raise when this
happens and no existing error.   We now set the userFlags when
creating an ATR, but this wasn't sufficient to insure the body is
actually binary.   Turns out we use SetDoc in a MutateInAsync call
to create it.   The MutateIn, under the covers, will use a transcoder
even when the value we gave the Spec is a byte array.   So... that
forced it to be a json doc no matter what.   The change here is if
you give a MutateInSpec a value which is a byte array, it is written
directly without any transcoding.   Other sdks do this, and it seems
to make sense.

Results
=======
All fit tests pass except the known issue with CustomSerializerTests.
Note there are some intermittant fails though - this will be looked
at along with any CI discrepancies shortly.

Change-Id: I3ec8bae904c39fbbc50e27d424537f000b175f9c
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/232753
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/235382
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>