NCBC-4054: Stellar gateway port is hardcoded
Motivation
==========
We may have the gateway on a different port (like, cbdino).  If so
we will specify that in the connection string, so we should use that
port.

Modification
============
If a port is specified, use it.   Otherwise use the default.

Results
=======
I can run fit tests with a cbdino-created stellar cluster.

Change-Id: I1cf664c9cd59f0e53eafad78ab53ee1e2ceb24cd
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/233314
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>