NCBC-4056: QueryOptionsTest issues
Motivation
==========
Fix a few issues with the fit QueryParametersTest for Stellar
connections.

Modification
============
Send the named and positional parameters across properly - they just
needed to be included.  The profile should not appear in the result
when excluded by the options, and we had a small parsing issue with
the times in the metrics.

Results
=======
QueryOptionsTest passes now

Change-Id: Ia33f9dfe2b051e76ac30f97b99304596099cd370
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/233531
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/235384
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>