NCBC-4057: ExpiryTest issues with Stellar
Motivation
==========
Stellar should have green tests in fit...

Modification
============
This had mostly to do with the proper setting of ExpiryTime and/or
ExpirySecs in the grpc.   Note that you can mix preserveTtl and
expiry when doing an upsert, so there's a bit of complication there
but simple enough.   Now a single function does the work, with
comments.

Also - added projections to the grpc message for gets.

Results
=======
All expiry tests pass in fit.

Change-Id: I35eee1ca69785990e24f3c222500d256a507bc3b
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/233789
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>