NCBC-4058: LookupIn<XXX>Test issues
Motivation
==========
Moving towards completing protostellar support...

Modification
============
This amounted to properly handling some error conditions in various
LookupIn scenarios.   Which means looking at the Status.Code in
the Spec and raising the correct exception.

Results
=======
LookupIn<XXX>Tests all pass, with the exception of one that is a
serializer issue.

Change-Id: Ica9068eecbd384923a0622801481203f2b732703
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/234012
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>