NCBC-4055 Correct Stellar exception when locked/unlocked
Motivation
==========
Make the sdk and stellar api match - both should throw the same
exception when discovering a document is locked, or trying to
unlock a locked doc.

Modification
=============
No more DocumentLockedException when stellar discovers an error with
the matching retry string, instead an UnambiguousTimeoutException is
thrown, to match the existing api behavior.

Results
=======
Touch/Unlock/DocNotLocked tests all pass both with stellar and the
regular api.

Change-Id: I79ad48999235de5bb2086d45e8191a4c24cd74ee
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/234470
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>