NCBC-4055:  [Stellar] Lock/Unlock/Touch issues
Motivation
==========
Correct the serialization of the error context in CouchbaseException.
This was partially corrected in a previous commit, but didn't do
a good enough job.   We also needed other fields from the ErrorContext,
for other tests.

Also - IErrorContext defined the interface with non-nullable
properties.  The derived classes all implemented it with nullable
properties - this builds until you add in the converter, and then
the complaints start.   Simple fix.

Modification
============
A new STJ Converter was created to ensure we serialize the fields
in a derived class when holding a reference to an interface.  This
was the underlying issue with some tests not having all the info
in the context that was expected.

Also - IErrorContext defined the interface with non-nullable
properties.  The derived classes all implemented it with nullable
properties - this builds until you add in the converter, and then
the complaints start.   Simple fix.

Results
=======
All the effected tests pass now.

Change-Id: Ib289b44214582b23a0a96966eb057cbeaf052229
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/234561
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/235390
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>