NCBC-4061: [Stellar] QueryIndexManager fails in FIT
Motivation
==========
Get all fit tests passing for protostellar.

Modification
============
The code that we generated from the protostellar repo didn't have
the WaitForIndexOnline code.   This modifies the CodeGen project
to be able to generate the code from the protostellar repo without
the need to reference a status.proto file (from google).   The
code was then re-generated using a later sha which had the
support for WaitForIndexOnline.   And then that call is now
supported in the QueryIndexManager for stellar.  We call the new
rpc call once per index name we are supplied with, in parallel.

Results
=======
Passing tests for QueryIndexManager in FIT.

Change-Id: Ieabfadcb61b01f3b3bb250f6df4ac763575f4325
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/234856
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/235391
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>