NCBC-4069 Serializer issues in stellar FIT tests
Motivation
==========
Lets get green fit tests.

Modification
============
When we set no transcoder in a LookupInAsync call, we are using the
DefaultTranscoder and DefaultSerializer.   In the LookupInGetTest
suite, the contentAsBytesGetLookup failed because we cannot de-
serialize the content using ContentAs<byte[]>.  This came down to
out use of the DefaultSerializer to Deserialize to a binary-friendly
type.   The specs say we should be able to do this - but it needed
to be done manually (just copying the bytes appropriately), as we'd
otherwise rely on Newtonsoft which will throw an exception.

While there, there is a workaround for this we added for transactions
in the LookupInResult#ContentAs<T> call, which we no longer need, so
it has been removed.  We needed to set the transcoder to BinaryTranscoder
and then insure any binary xattrs would use that transcoder, like
when we deserialize the contents of the document.   But that is now
unnecessary.   Later we will adjust the transactions code to not
bother with this but it works either way for now.

Results
=======
The contentAsBytesGetLookup passes.   The other test that seems to
have a serializer issue turns out to be covered by another but that
calls out our lack of use of the passed in Transcoder in Stellar, so
we will have green tests when that is done.

Change-Id: I3ad608e31830c159733113c9d53719e11ed58724
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/235028
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/235392
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>