NCBC-4023: Support Transcoders in Stellar
Motivation
==========
Transcoders should be supported as in the regular api.

Modification
============
Mostly the changes are just using transcoder from the options if
it is available, instead of the serializer or default serializer.
Added a TypeTranscoder property to the StellarCluster with a default
to use as the fallback.  Tried to mimic the serializer vs transcoder
decision we make when decoding specs (only use transcoder for the
whole document).

Results
=======
All FIT tests pass for Stellar.

Change-Id: I37ce79f987c49e6d2458c058cbd8465c0edadbfe
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/235326
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>