NCBC-4079: Initial support for Expiry in Transactions
Motivation
==========
We should preserve the expiry of any document that has one, when
it is involved in transactions.   We cannot do this in query mode,
but in KV mode we can.   Also, we should be able to insert or
replace and give the document an expiry, as long as we are not in
query mode.

Modification
============
All mutations now use PreserveTtl if it is supported by the server.
Insert and Replace options include an expiry, and it is added to
the document when unstaged.  If an expiry is used when in query mode,
we fail immediately with a FeatureNotAvailable, as this is a logical
error in the transaction lambda.  Similarly, when there is an expiry
staged, we error out if attempting to switch to query mode.   We can
apply the expiry when cleaning up, on the rare occasions where we do
mutate the document when cleaning.

Results
=======
Passing the initial TTL fit tests.

Change-Id: I84ec7f4a800f8d54a46c495bf7722bc2d1147a18
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/235730
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>