NCBC-4073: Nightly builds on Github
Motivation
==========
Customers wanting the latest _before_ it is officially supported,
can find that on the github nuget repo.

Modification
============
Need to have the github url in the metadata, so this can be shown
in the github ui, under packages, and in the Nuget UI too.

Results
=======
Should make them show up on in the Packages section of the github
UI for couchbase-net-client.

Change-Id: I6277fe886f88c0fca35071d29371ae0e15023d46
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/236326
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>