NCBC-4113 Recent regressions in transactions fit tests
Motivation
==========
I see FIT test fails that are new - lets fix them.

Modification
============
Seems that when we use .ReviveDocument(true) unstaging an insert
or a replace after an insert in the same transaction, we would
sometimes _also_ have a PreserveTtl(true) added.   Yikes - that
is only appropriate when the doc isn't a tombstone.   Simple fix.

Result
======
This in combination with a fix in the performer result in no new
FIT test issues.

Change-Id: Ie97408f04f6116d0f0a797a5b25f0de2be02fce3
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/237936
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>